requirements.txt
agno
openai
duckduckgo-search
python-dotenv
yfinance
pypdf
requests
toon-format
tiktoken
